fjern Adresse-komponent & erstatt lignende logikk i endre-mottaker & friends
